feat: OutputView -> ResultView, TicketsView 클래스 분리
